Refactor stock and quantity fields to use Long type
Changed stock and quantity fields from BigDecimal/Double to Long across DTOs, entities, and services for consistency and improved type safety. Added soldQuantity tracking to ProductEntity and ProductResponse, updated mappers and services to handle sold quantity, and enhanced product filtering with new options for product/category search and sorting by sold quantity.

+++ src/main/java/org/stockify/dto/request/product/ProductCSVRequest.java
@@ -27,7 +27,7 @@ public class ProductCSVRequest {
 
     @Schema(description = "Stock quantity available", example = "50")
     @CsvBindByName(column = "stock")
-    private BigDecimal stock;
+    private Long stock;
 
     @Schema(description = "SKU (Stock Keeping Unit) code", example = "PHONE-001")
     @CsvBindByName(column = "sku")
